requirements.txt
# LiveKit Agents Framework
livekit-agents~=1.0

# AI Provider Plugins (add as needed)
livekit-plugins-openai~=1.0
livekit-plugins-deepgram~=1.0
livekit-plugins-silero~=1.0
livekit-plugins-turn-detector~=1.0

# Optional plugins (uncomment as needed)
# livekit-plugins-cartesia~=1.0
# livekit-plugins-elevenlabs~=1.0
# livekit-agents[mcp]~=1.0

# Development dependencies
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
